Refactored CCInertial class
