+++ CCLocation/Inertial/CCInertial.swift
@@ -10,16 +10,6 @@ import Foundation
 import CoreMotion
 import ReSwift
 
-struct PedometerData {
-    let endDate: Date
-    let numberOfSteps: Int
-}
-
-struct YawData {
-    let yaw: Double
-    let date: Date
-}
-
 protocol CCInertialDelegate: class {
     func receivedStep(date: Date, angle: Double)
 }
@@ -34,7 +24,7 @@ class CCInertial: NSObject {
     private var previousPedometerData: PedometerData?
     private var yawDataBuffer: [YawData] = []
     
-    fileprivate var currentInertialState: InertialState!
+    var currentInertialState: InertialState!
     
     weak var stateStore: Store<LibraryState>!
     
@@ -46,6 +36,33 @@ class CCInertial: NSObject {
         self.stateStore = stateStore
         currentInertialState = InertialState(isEnabled: false, interval: 0)
         stateStore.subscribe(self)
+    }
+    
+    internal func start() {
+       if #available(iOS 11.0, *) {
+           let pedometerAuthStatus = CMPedometer.authorizationStatus()
+           
+           if pedometerAuthStatus == .authorized || pedometerAuthStatus == .notDetermined {
+               Log.debug("Starting intertial")
+               
+               startCountingSteps()
+               startMotionUpdates()
+           } else {
+               Log.debug("Authorisation status is .denied or .restriced, no inertial updates requested")
+           }
+       } else {
+           Log.debug("Starting intertial")
+           
+           startCountingSteps()
+           startMotionUpdates()
+       }
+    }
+       
+    internal func stop() {
+        Log.debug("Stopping inertial")
+           
+        pedometer.stopUpdates()
+        motion.stopDeviceMotionUpdates()
     }
     
     internal func startCountingSteps() {
@@ -69,16 +86,13 @@ class CCInertial: NSObject {
     private func startMotionUpdates() {
         if !motion.isDeviceMotionAvailable { return }
         
-        // need to double check the behaviour on this one
-        // self.motion.showsDeviceMovementDisplay = true
-        
         self.motion.startDeviceMotionUpdates(using: .xArbitraryZVertical,
                                              to: OperationQueue.main,
                                              withHandler: { (deviceMotion, error) in
             guard let data = deviceMotion, error == nil else {
                 let cmError = error as? CMError
                 
-                Log.error("Received: \(cmError.debugDescription)")
+                Log.error("Received motion update error: \(cmError.debugDescription)")
                 return
             }
             
@@ -86,7 +100,8 @@ class CCInertial: NSObject {
             let yawData = YawData(yaw: yawValue, date: Date())
             
             Log.verbose("""
-                Pedometer: Yaw: \(yawData.yaw)
+                Pedometer Data
+                Yaw: \(yawData.yaw)
                 Timestamp: \(yawData.date.timeIntervalSince1970)
                 Interval: \(self.motion.deviceMotionUpdateInterval )
                 """)
@@ -102,15 +117,12 @@ class CCInertial: NSObject {
             }
         })
     }
-
-    private func setInterval(time: TimeInterval) {
-        self.motion.deviceMotionUpdateInterval = time
-    }
     
     private func findFirstSmallerYaw(yawArray: [YawData], timeInterval: TimeInterval) -> YawData? {
         for (index, yaw) in yawArray.reversed().enumerated() {
             Log.verbose("""
-                Pedometer: Yaw time interval: \(yaw.date.timeIntervalSince1970)
+                Pedometer Data
+                Yaw time interval: \(yaw.date.timeIntervalSince1970)
                 Time interval: \(timeInterval)
                 """)
             
@@ -123,36 +135,7 @@ class CCInertial: NSObject {
         return nil
     }
     
-    private func rad2deg(_ number: Double) -> Double {
-        return number * 180 / .pi
-    }
-    
-    internal func start() {
-        if #available(iOS 11.0, *) {
-            let pedometerAuthStatus = CMPedometer.authorizationStatus()
-            
-            if pedometerAuthStatus == .authorized || pedometerAuthStatus == .notDetermined {
-                Log.debug("Starting intertial")
-                
-                startCountingSteps()
-                startMotionUpdates()
-            } else {
-                Log.debug("Authorisation status is .denied or .restriced, no inertial updates requested")
-            }
-        } else {
-            Log.debug("Starting intertial")
-            
-            startCountingSteps()
-            startMotionUpdates()
-        }
-    }
-    
-    internal func stop() {
-        Log.debug("Stopping inertial")
-        
-        pedometer.stopUpdates()
-        motion.stopDeviceMotionUpdates()
-    }
+    //MARK: - Handle Updates
     
     private func handleStepsSinceLastCount(fromPedometerData pedometerData: CMPedometerData) {
         guard let tempPreviousPedometerData = previousPedometerData else {
@@ -169,7 +152,7 @@ class CCInertial: NSObject {
             let oneStepTimeInterval = TimeInterval(periodBetweenStepCounts / Double(stepsBetweenStepCounts))
             
             Log.debug("""
-                Pedometer:
+                Pedometer Data
                 Step count: \(numberOfSteps)
                 Previous step count: \(tempPreviousPedometerData.numberOfSteps)
                 Period between step counts: \(periodBetweenStepCounts)
@@ -191,6 +174,7 @@ class CCInertial: NSObject {
                                           oneStepTimeInterval: TimeInterval,
                                           previousPedometerData: PedometerData) {
         Log.debug("Handle \(totalSteps) steps")
+        
         for i in  1 ... totalSteps {
             let tempTimePeriod = TimeInterval(Double(i) * oneStepTimeInterval)
             let tempTimeStamp = previousPedometerData.endDate.timeIntervalSince1970 + tempTimePeriod
@@ -201,7 +185,8 @@ class CCInertial: NSObject {
             }
             
             Log.debug ("""
-                Pedometer: Step count: \(i)
+                Pedometer Data
+                Step count: \(i)
                 Time period: \(tempTimePeriod)
                 Timestamp: \(tempTimeStamp)
                 Yaw value: \(String(describing: tempYaw.yaw))
@@ -213,40 +198,17 @@ class CCInertial: NSObject {
         }
     }
     
+    // MARK: - Helpers
+    
+    func setInterval(time: TimeInterval) {
+        self.motion.deviceMotionUpdateInterval = time
+    }
+    
+    private func rad2deg(_ number: Double) -> Double {
+        return number * 180 / .pi
+    }
+    
     private func getPeriodBetween(recentDate date1: Date, oldDate date2: Date) -> TimeInterval {
         return date1.timeIntervalSince1970 - date2.timeIntervalSince1970
-    }
-}
-
-// MARK:- StoreSubscriber delegate
-extension CCInertial: StoreSubscriber {
-    
-    public func newState(state: LibraryState) {
-        guard let newInertialState = state.intertialState else {
-            return
-        }
-        
-        if newInertialState != currentInertialState {
-            if let interval = newInertialState.interval {
-                setInterval(time: Double(interval) / 1000)
-            }
-            Log.debug("Pedometer: New state is: \(newInertialState)")
-            
-            updateCurrentInertialStateActivity(newState: newInertialState)
-            currentInertialState = newInertialState
-        }
-    }
-    
-    private func updateCurrentInertialStateActivity(newState: InertialState) {
-        guard let isInertialEnabled = newState.isEnabled else {
-            return
-        }
-        
-        if isInertialEnabled && !self.currentInertialState.isEnabled! {
-            start()
-        }
-        if !isInertialEnabled && self.currentInertialState.isEnabled! {
-            stop()
-        }
     }
 }
