+++ CCLocation/Inertial/Inertial+Structures.swift
@@ -1,0 +1,19 @@
+//
+//  Inertial+Structures.swift
+//  CCLocation
+//
+//  Created by Mobile Developer on 22/08/2019.
+//  Copyright © 2019 Crowd Connected. All rights reserved.
+//
+
+import Foundation
+
+struct PedometerData {
+    let endDate: Date
+    let numberOfSteps: Int
+}
+
+struct YawData {
+    let yaw: Double
+    let date: Date
+}
